Rename check to reportingAvailable
